requirements.txt
Pillow
psutil
requests
yagmail
keyboard
nudenet
pyinstaller
cryptography